Performance tests: Move measurement features from ReplaySH to PerfTestSH
